feat(remix): migrate Audiences examples to Segments API
