+++ remix-resend-examples/javascript/app/routes/api.segments.contacts.js
@@ -4,16 +4,16 @@ import { Resend } from "resend";
 const resend = new Resend(process.env.RESEND_API_KEY);
 
 export async function loader() {
-  const audienceId = process.env.RESEND_AUDIENCE_ID;
+  const segmentId = process.env.RESEND_SEGMENT_ID;
 
-  if (!audienceId) {
+  if (!segmentId) {
     return json(
-      { error: "RESEND_AUDIENCE_ID not configured" },
+      { error: "RESEND_SEGMENT_ID not configured" },
       { status: 500 }
     );
   }
 
-  const { data, error } = await resend.contacts.list({ audienceId });
+  const { data, error } = await resend.contacts.list({ segmentId });
 
   if (error) {
     return json({ error: error.message }, { status: 500 });
