+++ remix-resend-examples/javascript/app/routes/segments.jsx
@@ -1,57 +1,55 @@
-import { useFetcher } from "@remix-run/react";
-import type { ActionFunctionArgs, LoaderFunctionArgs, MetaFunction } from "@remix-run/node";
+import { useFetcher, useLoaderData } from "@remix-run/react";
 import { json } from "@remix-run/node";
-import { useLoaderData } from "@remix-run/react";
 import { Resend } from "resend";
 import { PageHeader } from "~/components/PageHeader";
 import { ResultDisplay } from "~/components/ResultDisplay";
 
 const resend = new Resend(process.env.RESEND_API_KEY);
 
-export const meta: MetaFunction = () => {
-  return [{ title: "Audiences - Remix + Resend" }];
+export const meta = () => {
+  return [{ title: "Segments - Remix + Resend" }];
 };
 
-export async function loader(_args: LoaderFunctionArgs) {
-  const audienceId = process.env.RESEND_AUDIENCE_ID;
+export async function loader() {
+  const segmentId = process.env.RESEND_SEGMENT_ID;
 
-  if (!audienceId) {
-    return json({ contacts: [], error: "RESEND_AUDIENCE_ID not configured" });
+  if (!segmentId) {
+    return json({ contacts: [], error: "RESEND_SEGMENT_ID not configured" });
   }
 
-  const { data: audiences } = await resend.audiences.list();
-  const { data: contacts } = await resend.contacts.list({ audienceId });
+  const { data: segments } = await resend.segments.list();
+  const { data: contacts } = await resend.contacts.list({ segmentId });
 
   return json({
-    audiences: audiences?.data || [],
+    segments: segments?.data || [],
     contacts: contacts?.data || [],
   });
 }
 
-export async function action({ request }: ActionFunctionArgs) {
+export async function action({ request }) {
   const formData = await request.formData();
-  const intent = formData.get("intent") as string;
+  const intent = formData.get("intent");
 
-  const audienceId = process.env.RESEND_AUDIENCE_ID;
-  if (!audienceId) {
-    return json({ error: "RESEND_AUDIENCE_ID not configured" }, { status: 500 });
+  const segmentId = process.env.RESEND_SEGMENT_ID;
+  if (!segmentId) {
+    return json({ error: "RESEND_SEGMENT_ID not configured" }, { status: 500 });
   }
 
   if (intent === "create") {
-    const email = formData.get("email") as string;
-    const firstName = formData.get("firstName") as string;
-    const lastName = formData.get("lastName") as string;
+    const email = formData.get("email");
+    const firstName = formData.get("firstName");
+    const lastName = formData.get("lastName");
 
     if (!email) {
       return json({ error: "Missing required field: email" }, { status: 400 });
     }
 
     const { data, error } = await resend.contacts.create({
-      audienceId,
       email,
       firstName: firstName || undefined,
       lastName: lastName || undefined,
       unsubscribed: false,
+      segments: [{ id: segmentId }],
     });
 
     if (error) {
@@ -62,14 +60,13 @@ export async function action({ request }: ActionFunctionArgs) {
   }
 
   if (intent === "remove") {
-    const contactId = formData.get("contactId") as string;
+    const contactId = formData.get("contactId");
 
     if (!contactId) {
       return json({ error: "Missing required field: contactId" }, { status: 400 });
     }
 
     const { error } = await resend.contacts.remove({
-      audienceId,
       id: contactId,
     });
 
@@ -83,22 +80,22 @@ export async function action({ request }: ActionFunctionArgs) {
   return json({ error: "Unknown intent" }, { status: 400 });
 }
 
-export default function Audiences() {
-  const loaderData = useLoaderData<typeof loader>();
-  const fetcher = useFetcher<typeof action>();
+export default function Segments() {
+  const loaderData = useLoaderData();
+  const fetcher = useFetcher();
 
   return (
     <div>
       <PageHeader
-        title="Audiences"
-        description="Manage audiences and contacts. Set RESEND_AUDIENCE_ID in your .env file."
+        title="Segments"
+        description="Manage segments and contacts. Set RESEND_SEGMENT_ID in your .env file."
       />
 
       <div style={{ marginBottom: "32px" }}>
         <h2 style={{ fontSize: "20px", marginBottom: "12px" }}>Contacts</h2>
         {loaderData.contacts && loaderData.contacts.length > 0 ? (
           <ul style={{ listStyle: "none", padding: 0 }}>
-            {loaderData.contacts.map((contact: any) => (
+            {loaderData.contacts.map((contact) => (
               <li
                 key={contact.id}
                 style={{
@@ -187,21 +184,20 @@ export default function Audiences() {
         <h3 style={{ margin: "0 0 12px 0", fontSize: "14px" }}>API Code</h3>
         <pre style={preStyle}>{`// Create a contact
 const { data, error } = await resend.contacts.create({
-  audienceId: "aud_...",
   email: "jane@example.com",
   firstName: "Jane",
   lastName: "Doe",
   unsubscribed: false,
+  segments: [{ id: "seg_..." }],
 });
 
 // List contacts
 const { data: contacts } = await resend.contacts.list({
-  audienceId: "aud_...",
+  segmentId: "seg_...",
 });
 
 // Remove a contact
 await resend.contacts.remove({
-  audienceId: "aud_...",
   id: contact.id,
 });`}</pre>
       </div>
@@ -209,14 +205,14 @@ await resend.contacts.remove({
   );
 }
 
-const labelStyle: React.CSSProperties = {
+const labelStyle = {
   display: "block",
   marginBottom: "4px",
   fontSize: "14px",
   fontWeight: 500,
 };
 
-const inputStyle: React.CSSProperties = {
+const inputStyle = {
   width: "100%",
   padding: "8px 12px",
   border: "1px solid #d4d4d8",
@@ -225,7 +221,7 @@ const inputStyle: React.CSSProperties = {
   boxSizing: "border-box",
 };
 
-const buttonStyle: React.CSSProperties = {
+const buttonStyle = {
   padding: "10px 20px",
   backgroundColor: "#18181b",
   color: "#fff",
@@ -236,7 +232,7 @@ const buttonStyle: React.CSSProperties = {
   cursor: "pointer",
 };
 
-const codeBlockStyle: React.CSSProperties = {
+const codeBlockStyle = {
   marginTop: "32px",
   padding: "20px",
   backgroundColor: "#18181b",
@@ -244,7 +240,7 @@ const codeBlockStyle: React.CSSProperties = {
   color: "#e4e4e7",
 };
 
-const preStyle: React.CSSProperties = {
+const preStyle = {
   margin: 0,
   fontSize: "13px",
   whiteSpace: "pre-wrap",
