+++ remix-resend-examples/javascript/package.json
@@ -14,7 +14,7 @@
     "isbot": "^5.1.0",
     "react": "^18.3.0",
     "react-dom": "^18.3.0",
-    "resend": "^4.0.0",
+    "resend": "^6.24.0",
     "svix": "^1.44.0"
   },
   "devDependencies": {
